fits reader is lazy
The fits reader is now lazy, and reads out whole tiles when it can.
Previously it read a scanline at a time, and used WIO.

We read in planes and combine with a set of im_bandjoin().

Use an im_flipver() at the end to get rows in the right order.