started adding new CLI interface
Started hacking on the thing to let you call operations from the
command-line. The target syntax is something like:

	add a b c

where a, b, c are the required parameters. Optional params may be appended to
the operation name, eg.

	add{saturated=true} a b c

Required params which are subclasses of VipsObject may themselves have
required and optional params, eg.

	affine in.v out.v bilinear{12,sharp=true} 0 1 1 0 0 0