remove the version numbers from the .pc files
we used to have minor version numbers in the pkg-confiog files, eg:

pkg-config vips-7.28 --cflags

The idea was to make working with multiple installed versions (something
I do a lot at least) easier. But it was also annoying and made upgrading
harder for downstream packages. And the benefits were very marginal.