use openslide_detect_vendor() on OpenSlide 3.4.0
It is much more efficient than openslide_open().  It also fixes a bug:

If openslide_open() failed, we had no way of knowing whether it was an
interesting failure (for relevant slide formats) or an uninteresting
one (for a generic TIFF).  So, the is_a method would always return false
in this case.  This could cause unexpected results; for example, on MIRAX
slides, VIPS would open the .mrxs file itself (a JPEG thumbnail) and the
user would be left wondering where all their pixels went.  Now, if there
is an interesting failure, is_a will succeed but header/load will fail.