Add alpha_q(uality) webpsave option, fix Q init
This allows to control the fidelity of the alpha channel by allowing
webp to reduce the number of palette entries from 256 at alpha_q=100
to 2 at alpha_q=0. This is most useful with cutout pictures where
something like alpha_q=30 or around 8 levels is usually sufficient
for smooth outlines.

For some reason webp->Q was intiailized to 80 even though the Q option
and libwebp use 75 by default. This was never used, so it's just to
avoid confusion about what default is used.