Vendor python.m4 to fix wrong multiarch exec paths
The python.m4 that ships with automake hardcodes the library path for
platfrom dependent files to `lib`, which is wrong for multiarch systems.

Instead we import the `m4/python.m4` from pygobject, which does not have
this problem and also ensures we have the same behavior as pygobject
which we are depending on for pyvips8.

Due to this change some minor changes to how the `bootstrap.sh` script
clean up the `m4` folder where needed.

The python search order has been modified to prefer the default python
over python2 over python3.