round tiff pyr layer sizes up
Previously, we found the sizes of pyr layers by repeated /2 and truncation
to int. This had the effect of trimming pixels off the right and
bottom of the image.

This patch rounds up instead, so we never lose any pixels, and always
calculating the layer size by division from the base image size, so
rounding errors cannot accumulate.

Since this allows half pixels, we have extra code to stretch layers when
necessary by copying the right-most edge and bottom-most row of pixels.