larger strip size for pdfload
We had 128-pixel high strips for pdfload. This kept mem use very low,
but meant many calls to page_render. For a tall page with an image
background, we could take almost an hour to render a single page as the
background was repeatedly repainted.

Change the strip height to 5000. This will render most pages in a single
call, but still allow huge images.

See https://github.com/jcupitt/libvips/issues/681