fix up autofit again
A bit cleaner and simpler now, stops earlier if it can.

I experimented with other ways of picking the next guess, but none were
significantly better than picking the half-way point. The problem is
that autosize can be very non-linear. Line breaks appear and disappear
as you adjust DPI and it's possible to construct cases where it never
tries the correct solution.