Support for operating on file descriptors
Descriptor arguments are distinguished by /proc/self/fd
prefix (the side-effect is that opening files under /proc/self
is no longer possible, but that is not the biggest problem with
this PR).

Since the driver executable does not support specifying
the file type, this implementation uses ugly hack to grok it
from argument, e.g. "/proc/self/11.jpg".