Use O_TMPFILE when available
O_TMPFILE is definitive solution for problem of
temporary files on Linux. The file, created with this flag
does not get registered in containing directory, so when
the application terminates, they are automatically deleted in
race-free manner.

O_TMPFILE files do not have a name, when open() is called with
this flag, it uses the "pathname" argument as name of containing
directory.