threadset: stop idle threads after a regular timeout (#3253)
* Add `vips_semaphore_down_timeout` to public API

In preparation for the next commit.

* threadset: stop idle threads after a regular timeout

In line with the previous behaviour of GLib's threadpool.

* threadset: fix a small memory leak

* threadset: ensure idle threads are freed on exit

GLib threads are spawned with `ref_count == 2`, this reference is
decreased during thread exit and associated resources are freed
during `g_thread_join()`. However, idle threads that are stopped
after a regular timeout are not joined. To fix this, decrease the
reference count during `vips_threadset_add()` and increase it just
before joining the thread with `g_thread_join()`.

See:
https://gitlab.gnome.org/GNOME/glib/-/blob/2d5d990c6ac7ff45bcc4485cce95e87687d47595/glib/gthread.c#L522
https://gitlab.gnome.org/GNOME/glib/-/blob/2d5d990c6ac7ff45bcc4485cce95e87687d47595/glib/gthread-posix.c#L1287

* Deprecate `vips_g_thread_join()` in favor of `g_thread_join()`

* nit: remove extra semicolon