Add fast path to extract_band and bandjoin for uchar (#3302)
This image format is common and these functions are used
by many other operations. Removing the inner loop allows
the compiler to auto-vectorise the remaining loop.

Extracting 1 channel or joining 1 and/or 3 channels
(e.g. RGB+A) is up to ~60% faster.

There is already a fast path for bandjoin_const
and this approach keeps things consistent.