colour: add AVX-specific XYZ2Lab implementation, ~10% faster
This conversion is most commonly used for colour difference
calculations and is often their bottleneck. It is also used
for the smartcrop and sharpen operations. ICC profile
transforms can use LAB as the connection space.

This adds up to ~4KB to the binary size.

Also moves the GOnce status to live in same scope as the LUT
it protects, which provides a small speed-up as its status
is no longer checked for every pixel.