Ensure libheif symbols are found when built statically
libheif is a C++ library that exposes a C API. When it's built
statically, the linker must also link with the C++ stdlib, which
is done automatically when using a C++ compiler.

Prior to commit strukturag/libheif@27097a6, libheif included
libstdc++ in the `Libs.private` field of their .pc file. This meant
it was exposed in the case of static linking, enabling a build
system for a C application to link against libstdc++ as needed.
However, this was not portable as libstdc++ is specific to the GNU
toolchain, when using Clang, you can optionally use libc++.

This commit aims to find symbols in libheif using a C++ compiler,
ensuring that Meson can confirm their availability at link-time,
even when libheif is statically built.