cpp: Fix build in Qt-using projects without changes
Qt may define the "signals" keyword, which clashes with C/GLib use of
the same term as a variable name.
With this change we temporarily disable the keyword to allow GLib
includes to work, before resetting it if Qt usage was detected.

This yields no unexpected surprises when using libvips from a Qt-using
C++ project.