Use release version for C++ shared library SONAME
Unlike the C ABI, the C++ ABI is not guaranteed to be compatible
between releases so use of the C ABI version for the C++ shared
library's SONAME is insufficient.

Where possible (i.e. not macOS), the C ABI version (currently 42)
is retained as a prefix to ensure backwards compatibility when
ordering against previous versions.

The project version is used as the version suffix to ensure
each release has a distinct SONAME.