vips7compat.h: remove `OrcProgram` and `OrcExecutor` stubs (#4207)
Remove the confusing `OrcProgram` and `OrcExecutor` stubs to avoid
conflicts with `orc.h`. Should be safe, the deprecated `VipsVector`
and `VipsExecutor` APIs was tightly coupled to liborc and became
no-op after the adoption of Highway.

While this is a potential ABI break, this API was likely not used
outside libvips. The same ABI break could also occur in versions
prior to 8.15 when a binary was built with `-Dorc=disabled` instead
of `-Dorc=enabled`.