morph: sync C-paths with the Highway implementation
Previously, `seq->coff` was used both for storing offsets to clear
values (zero values in masks) and as an array for non-128 mask
coefficients. However, in commit 40e2884 (PR #3618), `seq->coff`
was restricted to `guint8` values, making it incompatible for
storing offsets.

Fix this by syncing the C-paths with the Highway implementation.