improve operation cache touch
On a operation cache hit, we were only touching the hit itself.

As a result, after some operations on a loaded image, the image itself
could be forced out by the operations that had been performed on it,
even though the underlying image was still being used.

This PR makes cache touches also update the timestamps on all upstream
operations, since the upstream operations are also being used
(indirectly).