WIP: experiment with gi-docgen (#3056)
* sort-of works

* fix some small issues

* more hacking

* better version numbers

* oops, dropped a change

* fix links in binding.md

and enable external linking

* rewrite `extending.xml` in markdown

* rewrite using-C in markdown

* rewrite file-format in markdown

* move function-list to markdown

and unify formatting

* ready to rename links

* revise

* revise patterns again

* generate rename list for methods

* add some missing (method) annotations

* more markup revision

* almost there

* just a few broken links now

* more enum expansions in rename

* oop missing comma in gen-function-list

* Fix some rebase issues

* Fix metadata

* Clarify comment

* doc: revise and port to gi-docgen linking syntax

* doc: move the buf SECTION

Move it to the VipsBuf struct docs.

* target.c: port to gi-docgen linking syntax

* doc: move the header SECTION

Move it to a separate page as there's no public struct to hang the
docs off.

* header.c: port to gi-docgen linking syntax

* Ensure GIR file includes the whole public API

It was missing public API functions such as `vips_sink_disc()`.

* Fix GIR warnings after bd4f4bcb

* doc: move the VipsImage SECTION

Move it to the struct docs.

* object: ensure all `VIPS_ARG_*` function macros are introspectable

* memory: ignore autocleanups when generating introspection data

* Ignore private/test API when generating introspection data

* Remove redundant gtk-doc comments

* Remove gtk-doc workaround

* doc: move the VipsRegion SECTION

Move it to the struct docs.

* doc: move the VipsOperation SECTION

Move it to the struct docs.

* doc: move the VipsObject SECTION

Move it to the struct docs. Also, separate VipsArgument to its own
struct doc.

* doc: move the VipsThreadState SECTION

Move it to the struct docs.

* doc: move the VipsForeign SECTION

Move it to the struct docs. Also, separate VipsForeignLoad and
VipsForeignSave to its own struct doc.

* doc: drop the rectangle SECTION

* doc: move the connection SECTION

Move it to the struct docs.

* doc: move the VipsSbuf SECTION

Move it to the struct docs.

* doc: move the interpolator SECTION

Move it to the struct docs. Also, port the whole file to gi-docgen
linking syntax

* Add docs for `Source`, `SourceCustom`, `Target` and `TargetCustom`

* doc: move the arithmetic SECTION

Move it to a separate page as there's no public struct to hang the
docs off.

* fix links in function-list.md

* add more methods to rename

* turn more func into method

* Appease clang-format

---------

Co-authored-by: Kleis Auke Wolthuizen <github@kleisauke.nl>