openslideconnection: simplify locking
_ref and _unref already need to happen inside the global lock, so we can
protect the refcount with the global lock and use a per-connection lock
only to protect the slow initialization of connection->osr.

That does mean that _free needs to openslide_close() inside the osr lock,
since otherwise there's a codepath (the _open revalidate case) where the
current thread might not have received an up-to-date ->osr.