openslideconnection: fixes and some cleanups (#4455)
* openslideconnection: fix unused-variable warning with assertions disabled

* openslideconnection: only openslide_set_cache() right after openslide_open()

There's no need to do it when reusing an existing openslide_t, and doing
so invalidates all the existing cache entries for the handle.

* openslideconnection: remember to take global lock around unref

_unref can call _free, which needs the lock.

* openslideconnection: use a GQueue to track unused connections

Eliminate some code, and potentially improve trim performance with a
large number of active connections.

* openslideconnection: simplify locking

_ref and _unref already need to happen inside the global lock, so we can
protect the refcount with the global lock and use a per-connection lock
only to protect the slow initialization of connection->osr.

That does mean that _free needs to openslide_close() inside the osr lock,
since otherwise there's a codepath (the _open revalidate case) where the
current thread might not have received an up-to-date ->osr.