fix linear resize of 16-bit rgba (#4467)
* fix linear resize of 16-bit rgba

We were changing the interpretation in icc_import and icc_export, but we
were not rescaling the alpha. As a result, 16-bit RGBA images had a
alpha of 65535 in PCS, not 255.

This PR moves alpha channel scaling into the base VipsColour class, so we
get correct scaling automatically, and moves XYZ <-> scRGB conversion
back on top of VipsColourCode, partially reverting #3627

see https://github.com/libvips/libvips/issues/4343

* Update libvips/colour/colour.c

Co-authored-by: Kleis Auke Wolthuizen <github@kleisauke.nl>

* review comments

---------

Co-authored-by: Kleis Auke Wolthuizen <github@kleisauke.nl>