target: avoid seek on bad file descriptor during TIFF save
After PR #4482, saving a TIFF file could trigger a seek on a bad
file descriptor. This occurs because the file is now lazily created
on the first write, and a seek may happen before that point.

This change ensures the file is also created on the first seek, if
necessary. Note that while most systems tolerate seeks on bad file
descriptors, this causes a runtime error on Windows when linked
against UCRT.